Move PointLight to the primitives package.
