+++ primitives/material.go
@@ -4,7 +4,6 @@ import (
 	"math"
 
 	"github.com/anolson/rtc/color"
-	"github.com/anolson/rtc/light"
 	"github.com/anolson/rtc/tuple"
 )
 
@@ -37,7 +36,7 @@ func DefaultMaterial() *Material {
 
 // Lighting shades an objects so it appears three-dimensional
 func (m *Material) Lighting(
-	light *light.PointLight,
+	light *PointLight,
 	point *tuple.Tuple,
 	eyev *tuple.Tuple,
 	normalv *tuple.Tuple,
